Feat: usecase用のテスト用関数を整理し適用
