+++ internal/usecase/usecasetest/README.md
@@ -1,35 +1,42 @@
 # internal/usecase/usecasetest
 
-概要: ユースケース層（ビジネスロジック）のテストを支援するためのユーティリティ群です。テスト用のインスタンス作成、依存コンポーネントの差し替え、共通セットアップ／クリーンアップ処理などを提供します。
+概要: ユースケース層のテストを支援するヘルパーを提供するパッケージです。本ディレクトリには、ユースケースの単体テストで頻繁に必要となるテスト用エラー生成やモック化されたトランザクションマネージャーのユーティリティが含まれます。
 
-## 役割
+## 提供する主な機能
 
-- ユースケースの単体／統合テストで使う共通セットアップ（依存注入、モック準備、コンテキスト生成など）を提供する。
-- テスト用のインスタンス（テストダブルや最小構成の実体）を組み立てるヘルパーを提供する。
-- テスト実行前後の初期化・破棄処理（リソースの解放、状態リセット）をまとめることで、テストの可読性と保守性を向上させる。
+- `ExpectedDBError(t *testing.T) error`
+  - テストで「DBエラーを表す固定のエラー」を簡単に生成するためのヘルパーです。テストの期待値として使います。
 
-## 必要度
+- `NewMockTransactionManager(t *testing.T) tx.Manager`
+  - gomock を用いて `tx.Manager` のモックを生成します。返却されるモックは `Do(ctx, fn)` を呼ばれた際に `fn(ctx)` をそのまま実行する振る舞いを持ち、`.AnyTimes()` により何度でも使えるよう設定されています。ユースケースのトランザクションまわりのロジックをテストする際に便利です。
 
-### 本番運用での必須度
+## 使い方（例）
 
-- 必須度: 不要
+1) DBエラーの期待値を作る
 
-理由: 本番コードにはテストユーティリティは含めず、実行環境には不要です。
+```go
+func Test_SomeUsecase_DBError(t *testing.T) {
+  expected := usecasetest.ExpectedDBError(t)
+  // expected をモックの返却値や期待エラー判定に利用
+}
+```
 
-### 開発/テスト運用での必須度
+1) トランザクションマネージャーのモックを使う
 
-- 必須度: 開発/テスト運用で必須
+```go
+func Test_SomeUsecase_WithTx(t *testing.T) {
+  mockTx := usecasetest.NewMockTransactionManager(t)
+  // mockTx をユースケースの依存に注入してテストを実行
+}
+```
 
-理由: ユースケース層のロジックはビジネスに直結するため、単体テストや統合テストで確実に検証する必要があります。本ディレクトリが提供するテストヘルパーはテストの安定化と簡素化に寄与します。
-
-### 無効化した場合の影響
-
-- ユースケースのテストにおいてセットアップ/ティアダウンが各テストに分散し、重複コードが増える。
-- モックやテストインスタンスの組み立てが煩雑になり、テストの信頼性や実行速度が低下する可能性がある。
+このモックは `Do(ctx, fn)` が呼ばれると `fn(ctx)` をその場で実行するため、トランザクション内の処理が呼ばれることを確認できます。
 
 ## 注意点
 
-- テスト用ユーティリティが本番リソース（外部サービス、本番DBなど）に誤って接続しないよう、環境変数や接続先を明確に切り分けてください。
-- 並列実行するテストでは共有状態による競合を避けるため、各テストで独立したモック/インスタンスを使うか、適切に同期をとってください。
-- テストで作成した外部リソース（ファイル、DBテーブル、ネットワークリソース等）はテスト終了時に必ず破棄してください。
-- テスト用ヘルパーは可能な限り副作用を抑え、テストの再現性を保つよう実装してください。
+- `NewMockTransactionManager` は内部で gomock を利用します。テスト内で生成したモックコントローラはテスト終了時に自動的に検証されます。
+- 本パッケージのユーティリティはテスト補助のためのものです。本番コードに組み込むべきではありません。
+
+## テスト
+
+- 本パッケージ自体にも単体テストが含まれている場合があります。`go test ./internal/usecase/usecasetest` で実行できます。
